🚚 move navbar component
