+++ frontend/lib/components/NavBar/components/Header/hooks/useHeader.ts
