+++ frontend/lib/components/NavBar/components/Logo.tsx
