+++ frontend/lib/components/NavBar/index.tsx
@@ -4,7 +4,7 @@ import { FC } from "react";
 import { Header } from "./components/Header";
 import { Logo } from "./components/Logo";
 import { MobileMenu } from "./components/MobileMenu";
-import { NavItems } from "./components/NavItems/";
+import { NavItems } from "./components/NavItems";
 
 export const NavBar: FC = () => {
   return (
